requirements.txt
agno==1.6.4
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
docstring_parser==0.16
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
google-genai==1.21.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
markdown-it-py==3.0.0
mdurl==0.1.2
openai
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
python-telegram-bot==22.1
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
rsa==4.9.1
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
telegram==0.0.1
tenacity==8.5.0
tomli==2.2.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
websockets==15.0.1
fastapi
uvicorn
exa-py
psycopg[binary]
tavily-python
